+++ functime/seasonality/fourier.py
